+++ Source/driver/Castro_setup.cpp
@@ -8,6 +8,7 @@
 #include <Castro_bc_fill_nd.H>
 #include <Castro_generic_fill.H>
 #include <Derive.H>
+#include <runtime_parameters.H>
 #ifdef RADIATION
 #include <Radiation.H>
 #include <RadDerive.H>
@@ -20,6 +21,9 @@
 #include <microphysics_F.H>
 #endif
 #include <eos.H>
+#ifdef NSE_THERMO
+#include <nse.H>
+#endif
 #include <ambient.H>
 
 using std::string;
@@ -208,19 +212,6 @@ Castro::variableSetUp ()
 
   init_prob_parameters();
 
-  // check to make sure that we didn't set any parameters that don't
-  // exist in C++ (like because of misspelling).  All of the problem.*
-  // parameters should have been accessed via parmparse at this point.
-
-  if (ParmParse::hasUnusedInputs("problem")) {
-      amrex::Print() << "Warning: the following problem.* parameters are ignored\n";
-      auto unused = ParmParse::getUnusedInputs("problem"); 
-      for (auto p: unused) {
-          amrex::Print() << p << "\n";
-      }
-      amrex::Print() << std::endl;
-  }
-
   // Initialize the runtime parameters for any of the external
   // microphysics (these are the parameters that are in the &extern
   // block of the probin file)
@@ -229,19 +220,19 @@ Castro::variableSetUp ()
   // set small positive values of the "small" quantities if they are
   // negative
   if (small_dens < 0.0_rt) {
-    small_dens = 1.e-200_rt;
+    small_dens = 1.e-100_rt;
   }
 
   if (small_temp < 0.0_rt) {
-    small_temp = 1.e-200_rt;
+    small_temp = 1.e-100_rt;
   }
 
   if (small_pres < 0.0_rt) {
-    small_pres = 1.e-200_rt;
+    small_pres = 1.e-100_rt;
   }
 
   if (small_ener < 0.0_rt) {
-    small_ener = 1.e-200_rt;
+    small_ener = 1.e-100_rt;
   }
 
 #if !defined(NETWORK_HAS_CXX_IMPLEMENTATION)
@@ -270,6 +261,29 @@ Castro::variableSetUp ()
   Real new_min_rho = std::max(small_dens, EOSData::mindens);
   small_dens = new_min_rho;
   EOSData::mindens = new_min_rho;
+
+  // Given small_temp and small_dens, compute small_pres
+  // and small_ener, assuming a more restrictive value is
+  // not already provided by the user. We'll arbitrarily
+  // set the mass fraction for this call, since we presumably
+  // don't need to be too accurate, we just need to set a
+  // reasonable floor.
+
+  eos_t eos_state;
+
+  eos_state.rho = castro::small_dens;
+  eos_state.T = castro::small_temp;
+  for (int n = 0; n < NumSpec; ++n) {
+      eos_state.xn[n] = 1.0_rt / NumSpec;
+  }
+#ifdef NSE_THERMO
+  set_nse_aux_from_X(eos_state);
+#endif
+
+  eos(eos_input_rt, eos_state);
+
+  castro::small_pres = amrex::max(castro::small_pres, eos_state.p);
+  castro::small_ener = amrex::max(castro::small_ener, eos_state.e);
 
   // some consistency checks on the parameters
 #ifdef REACTIONS